Create gameplay loop and functions
Updated the JS code to end the game once either player reaches a score
of 5 points. Then, the page displays the winner and prompts the player
to select a button to play again, doing so resets the score and lets
the player play another game. Also cleaned up the JS code by removing
more unnessecary functions and code that was leftover from the original
version of the game.